+++ integrations/kecho/echo-v4.go
@@ -1,4 +1,4 @@
-package Kecho
+package kecho
 
 import (
 	"bytes"
