+++ integrations/kgin/gin-v1.go
@@ -1,4 +1,4 @@
-package Kgin
+package kgin
 
 import (
 	"bufio"
