+++ integrations/kwebgo/webgo-v4.go
@@ -1,4 +1,4 @@
-package Kwebgo
+package kwebgo
 
 import (
 	"bytes"
